requirements.txt
gitpython
faiss-cpu
sentence-transformers
numpy
scikit-learn
torch
openai
accelerate
transformers
tiktoken